moved to use database instead of db
