+++ src/main/java/database/MemeDB2000.java
@@ -1,4 +1,4 @@
-package db;
+package database;
 
 import java.sql.*;
 import java.util.ArrayList;
