+++ src/main/java/database/MemeDBC2000.java
@@ -1,4 +1,4 @@
-package db;
+package database;
 
 import datastructures.MemeDBMsg2000;
