+++ src/test/java/database/MemeDBC2000Test.java
@@ -1,4 +1,4 @@
-package db;
+package database;
 
 import datastructures.MemeDBMsg2000;
 import org.junit.After;
